Fix time calculation bug in GameManager.py

+++ jhockey/GameManager.py
@@ -259,10 +259,10 @@ class GameManager:
                 self.state = GameState.STOPPED
 
             if self.broadcaster is not None:
-                time_left_decisecond = int(self.timer.get().total_seconds * 1e1) if self.timer.timestarted else self.match_length_sec * 1e1
+                time_left_decisecond = self.timer.get().total_seconds * 1e1 if self.timer.timestarted else self.match_length_sec * 1e1
                 self.broadcaster.set_message(
                     BroadcasterMessage(
-                        time=time_left_decisecond,
+                        time=int(time_left_decisecond),
                         robots=self.robot_states,
                         enabled=self.state == GameState.RUNNING,
                     )
